extract recordLastNum
